src/backend/wellness/pyproject.toml
[project]
name = "wellness"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.18.3",
    "asyncpg==0.31.0",
    "cashews[redis]==7.4.4",
    "celery[redis]>=5.6.2",
    "greenlet>=3.3.1",
    "grpcio==1.76.0",
    "grpcio-tools==1.76.0",
    "py-async-grpc-prometheus==0.2.0",
    "sqlalchemy==2.0.46",
    "uvicorn[standard]==0.40.0",
]

[dependency-groups]
test = [
    "coverage==7.13.3",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
]
